requirements.txt
fastapi
deepface
retina-face
pymongo
pillow
matplotlib
pandas
uvicorn
python-dotenv

